+++ R/utilities_caching.R
@@ -58,11 +58,13 @@ check_if_cache_update_needed <- function(function_name){
   df <- retrieve_cache(function_name)
   current_atlas <- potions::pour("atlas", "region", .pkg = "galah")
   # build some checks
+  caching_disabled <- !potions::pour("package", "caching", .pkg = "galah")
   is_local <- !is.null(attr(df, "ARCHIVED"))
   is_wrong_atlas <- attr(df, "region") != current_atlas
   is_too_short <- nrow(df) < 1 # somewhat arbitrary, but catches empty tibbles
   # evaluate those checks
-  result <- is_local | is_wrong_atlas | is_too_short # if any, update is needed
+  # if any are TRUE, update is needed
+  result <- is_local | is_wrong_atlas | is_too_short | caching_disabled
   if(length(result) < 1){ # bug catcher
     TRUE
   }else{